Bcc in appended msg, stripped out of sendMail-submission
The Bcc header is currently not saved in the message
appended to the sent-mail folder due to some SMTP servers
not stripping it out upon submission. This patch is the first
in a series that intends to do the stripping-out client-side.

Specifically, this patch saves the Bcc header in the local
message that is appended and strips out the Bcc header in the
local message before submitting using the sendMail method.

For future-proofing, the stripping approach is not coupled to
the Bcc header, so that it can be reused, e.g., for the
Resent-Bcc header when Resending is added to Trojita.

BUG: 356865
Change-Id: Iaed29f20c9f6d2aa34ebdbec662b921cd84a7f22