Do not modify filename of directly opened attachments
Attachments that are opened using “Open directly” are saved to a
temporary file in the OS-defined directory for temporary files.
They are currently saved there with a modified filename to ensure
uniqueness (to not overwrite other files there). This filename
modification is a bit inconvenient for a user that wants to save the
file after having opened it by this route, as they most likely will
remove the prepended string. This patch saves the file in a
directory with a unique name, without changing the name of the file
itself (or naming it ‘anonymous_file’ if the filename is empty),
thus removing the inconvenience.

Change-Id: I4d07c8e6fe105d6121f3f46b32600a253561cab9