Replace KIMAP-derived utf7-codec by more readable self-coded version
The KIMAP-derived encoder and decoder for utf7 is not very easy to
read. This patch replaces it by a self-coded version that is much easier
to read. It builds on QTextCodec for base64 and QRegularExpression
to identify segments that need to be transcoded.

Obviously, changing a piece of code that works well may be
contentious. However, this does (should) make the codebase more
accessible.

Change-Id: Ieda0063effe6cc9883806f1a191d7203aeebfadf