refactor response handling and case/if-else use in qwwsmtpclient
To be able to extend qwwsmtpclient (for, e.g., pipelining, chunking),
its structure needs to be accessible for current developers. To this
end, the response handling has been transitioned from a line-by-line
one to a batched one (this is is more appropriate in a pipelining
context). Because this impacts the way in which the command responses
are processed, the case/if-else tree that did this processing has
been heavily refactored. By replacing many if-else groups by switch
statements, this tree has become less deep and so more readable.

Change-Id: I09ed5cf2559fdd4aeef0e5734433c9ef4a86b672